Added a command oq plot_sites csvfiles
