+++ openquake/baselib/hdf5.py
@@ -24,6 +24,7 @@ import tempfile
 import warnings
 import importlib
 import itertools
+from dataclasses import dataclass
 from urllib.parse import quote_plus, unquote_plus
 import collections
 import json
@@ -41,8 +42,6 @@ vuint16 = h5py.special_dtype(vlen=numpy.uint16)
 vuint32 = h5py.special_dtype(vlen=numpy.uint32)
 vfloat32 = h5py.special_dtype(vlen=numpy.float32)
 vfloat64 = h5py.special_dtype(vlen=numpy.float64)
-
-CSVFile = collections.namedtuple('CSVFile', 'fname header fields size skip')
 FLOAT = (float, numpy.float32, numpy.float64)
 INT = (int, numpy.int32, numpy.uint32, numpy.int64, numpy.uint64)
 MAX_ROWS = 10_000_000
@@ -50,6 +49,18 @@ MAX_ROWS = 10_000_000
 if sys.platform == 'win32':
     # go back to the behavior before hdf5==1.12 i.e. h5py==3.4
     os.environ['HDF5_USE_FILE_LOCKING'] = 'FALSE'
+
+
+@dataclass
+class CSVFile:
+    fname: str
+    header: list[str]
+    fields: list[str]
+    size: int
+    skip: int
+
+    def read_df(self):
+        return pandas.read_csv(self.fname, skiprows=self.skip)
 
 
 def sanitize(value):
@@ -934,7 +945,8 @@ def sniff(fnames, sep=',', ignore=set()):
                 common = set(header)
             else:
                 common &= set(header)
-            files.append(CSVFile(fname, header, common, os.path.getsize(fname), skip))
+            files.append(
+                CSVFile(fname, header, common, os.path.getsize(fname), skip))
     common -= ignore
     assert common, 'There is no common header subset among %s' % fnames
     return files
